fix some mistakes and improve AMUtil#consumeItemsAndSpawnEntity
